Compute aggregate data in the background
